Migrate to Next.js app directory
This allows us to use new the server components feature, which means
less JS gets sent to clients and more is done on the server-side, which
feels like a nice easy win.

It's not a perfect migration, since Next.js app directory still has a
few rough edges. For instance, we have a `[...slug]` page that acts as a
polyfill for a proper catch-all 404 page.
