--- content/site.json
@@ -1,24 +1,0 @@
-{
-  "title": "Ross Zurowski",
-  "description": "Designer and developer from Toronto.",
-  "email": "ross@rosszurowski.com",
-  "url": "https://rosszurowski.com",
-  "socialLinks": [
-    {
-      "title": "Email",
-      "href": "mailto:ross@rosszurowski.com"
-    },
-    {
-      "title": "GitHub",
-      "href": "https://github.com/rosszurowski"
-    },
-    {
-      "title": "Twitter",
-      "href": "https://twitter.com/rosszurowski"
-    },
-    {
-      "title": "Are.na",
-      "href": "https://www.are.na/ross-zurowski"
-    }
-  ]
-}
