+++ next.config.js
@@ -10,6 +10,7 @@ const config = {
     formats: ["image/avif", "image/webp"],
   },
   experimental: {
+    appDir: true,
     legacyBrowsers: false,
   },
   async redirects() {
