+++ package.json
@@ -25,7 +25,7 @@
     "feed": "^4.2.2",
     "globby": "^13.1.2",
     "next-contentlayer": "^0.2.9",
-    "next-plausible": "^3.6.4",
+    "next-plausible": "^3.6.5",
     "postcss": "^8.4.14",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
