+++ src/app/head.tsx
@@ -1,0 +1,9 @@
+import MetaTags from "src/components/meta"
+
+export default function Head() {
+  return (
+    <>
+      <MetaTags />
+    </>
+  )
+}
