+++ src/app/not-found.tsx
@@ -1,0 +1,5 @@
+import NotFound from "src/components/not-found"
+
+export default function NotFoundPage() {
+  return <NotFound />
+}
